+++ python-framework/duyi-service/apps/web-service/test/integration/conftest.py
@@ -36,6 +36,9 @@ def _create_cur():
 
 
 def pytest_sessionstart(session):
+    if session.config.option.collectonly:
+        return
+
     with _create_cur() as cur:
         cur.execute(
             f"SELECT pg_terminate_backend(pg_stat_activity.pid) "
@@ -60,6 +63,9 @@ def pytest_sessionstart(session):
 
 
 def pytest_sessionfinish(session, exitstatus):
+    if session.config.option.collectonly:
+        return
+
     with _create_cur() as cur:
         cur.execute(
             f"SELECT pg_terminate_backend(pg_stat_activity.pid) "
